This commit is a refactor of the SketchModel and ModelManager classes.
SketchModel was completely rewritten with a new interface,
and the rest of the commit are the changes required to make
everything else build and pass the unit tests again.

SketchModel now holds data to keep track of multiple resolution
versions of each object and also multiple conformations (variations)
of each object.  A conformation is something like a mutated
version of the protein and should in some ways be considered the
same type of object since we may want objects to switch which
version of the object that they are using.  Each conformation
can have multiple resolutions.

The resolutions currently supported are defined in an enum,
ModelResolution::ResolutionType (in sketchmodel.h).  The new
SketchModel inherits from QObject to take advantage of signal/slot
connections for simplifying objects on a background thread and
sending the data to the SketchObject (which is in a foreground thread).

SketchModel now also rotates and translates its input data to a given configuration,
currently so that the PCA oriented bounding box is also the axis aligned
bounding box.  Eventually this may be used to register multiple
conformations of the same molecule.  This also makes grabbing objects
a bit better since this bounding box is used there.

Explicit tests for the new SketchModel code will be added in the
next commit.