Removed the positioning code from SketchModel.
In this commit, I also removed code from projecttoxml that handled
conversion from the old way of doing models to the new one.  The
reason for removing the positioning/orienting code in SketchModel
pointed out an aggravating corner case, so since I am still the only
one with projects I decided to break save compatability.

This commit also removes tests of the SketchModel functionality that
got taken out and updates the test of save and load with the changes
explained above.