Changed the File->Import->PDB ID menu item's behavior.
It now downloads the PDB and populates all levels of a model with
the correct resolution obj file.  I refactored a bit in subprocesses
to add a "composite" subprocess that runs several subprocesses one
after the other.  I also changed the Blender decimation code so that
there is an option to set the target number of triangles instead of
setting the target percentage of decimation.  The ModelFromPDBRunner
class still needs improving and all the subprocess stuff needs unit
tests.  Committing this since it is the first working version.