Added a filter to project a geometry onto an arbitrary plane.
The plane is defined as a point and normal vector.  The filter
works, according to the tests that are being added with it.

I may be fixing the VTK documentation on creating filters after
this.  Some of the method documentations were terrible.  I just
need someone to explain them to me first.