Added a camera-attached 'floor' plane to project shadows onto.
The plane may be bright red (so I could see it in testing), but
it is in the right place and the right size.

The other changes are cleanup of some unnecessary #includes in
sketchproject.h and transformmanager.h and the associated changes.
I'm cleaning these up as I have to change header files, so eventually
they will all be fixed.