Objects now have shadows, although on the wrong plane.
Objects have shadows, although they are on the y=0 plane
and not on the view-relative plane.  This commit also fixes
other things that I found while working on this:

There were some method declarations left in worldmanager.h
that no longer exist there.  They had been moved elsewhere,
so I removed the declarations.

WorldManager had an unused field and its constructor was taking
a parameter to set that field.  I removed the field and the
parameter and fixed the one place that creates a WorldManager.