Trackers now have shadows.
Important notes: trackers had to be scaled up 50 times
for their shadows to be visible (the visible tracker is
still the same size, but the shadow is large).

For errors with shadow plane or tracker shadows and tracker
shadow color see #defines in sketchproject.cpp.  To change
object shadow colors, see the #define at the top of worldmanager.cpp.

All these #defines should probably move to their own header file.