Restructured the project into subdirectores. ON MY OWN TIME
Note this was done on my own time, not as part of my hours.
This has been on my todo list for a while and it was bugging me.
Future commits on this branch will fix c++ coding style and practice
things that I did not know about and did wrong.  Things that don't
affect how the code runs.  The branch name says it all.