Feature686 Temporary files are in correct place
Temporary files used to be created in the current directory
of the program which would break under some operating sytstems/
conditions.  Temporary files are now created in the system
temporary folder or in the project folder (whichever is more
applicable to that case).

Removed a (now) unneeded print in shadows code.