BUGFIX - Couldn't find subprocess on Windows
This was a combination of using the wrong ifdef check to
test if the platform was windows and some compile errors
in the block that were never found since the block was never
actually compiled.  Then there was a logic error with the
setup of the file dialog...

But we can now find subprocesses on Windows and they will run.