testStructureReplicator now does what it sounds like
In preparation for some refactoring with the StructureReplicator,
I changed the test to actually test it.  That test file had become
a scratchpad for any test I wanted to run and did not test what it
was meant to.  I moved the current test to a location that made more
sense for it and made some tests of StructureReplicator.  These do not
fully test it... it needs at least one more test before it fully tests
the structure replicator.