Feature582 (progress) added initial Chimera extension
For this feature we decided to extend Chimera to export .vtk files
that contain the data we want.  This initial version creates and
knows how to print a datastructure that mimics the format of the
VTK file.  Since I have to write all points first then all trianges
then all data arrays, it made sense to me to make a datastructure
that will hold them all and first populate it, then write to the
file once all the data has been consolidated.  As of this commit
it knows how to print the datastructure, but doesn't know how to
populate from chimera's data yet.