Feature582 - exports atoms, bonds and surfaces
The chimera extension now knows how to put the data for
atom locations, bonds, and surface geometry into the datastructure
for export to vtk.  The only data array currently exported is a
model number to deal with the case where there are multiple models
open in Chimera.  The result can be thresholded by this to select
a specific model.