Feature582 (progress) removed hydrophobicity export
Since VTK cannot read NaN (not a number) in its input arrays due to
issues with the underlying std::istream, we are not supporting the
export of this array for now.  The data can be positive, negative,
zero or not there, and since NaN is unavailable to represent "not
there", we would have to use an in-band exception value.  This is
probably not a good idea, and after some discussion with Russ, I am
removing the array.