Feature582 (progress) Fixed chimera export script
Apparently my interpretation of point data vs. field data in VTK
was off.  Point data is the dataset that is currently being used to
color the model, field data is all the other datasets (or that is how
the vtkPolyDataReader was interpreting it.  Only the first data array
was being read in to SketchBio.  Note: Paraview has its own reader that
handled my bad files better than the default vtkPolyDataReader did.