Feature582 (progress) Modified the C++ code to use new chimera extension
This commit modifies the C++ code to expect a VTK file with data from
Chimera.  The code currently assumes that the extension is already in
Chimera, but can be modified to change Python's path and put it in temporarily.
It does not yet threshold out the atoms and bond which currently display along
with the surfaces.  It generates simpler surfaces using the vtkDecimatePro filter
instead of going out to Blender. (So much for background simplification).  I cannot
yet export the VTK data to Blender though.  That is the biggest thing that is left
on this feature.