Feature582 (progress) initial version of vrml writer
This is the initial version of a VRML writer that given a polydata and
a color map will write that data out to a file for import into Blender
or other modeling/animation tools.  Currently it exports geometry only,
color will be done later.  Polygons are the only type of cell that is
currently supported since they are easiest and (I think) the most common
in our data.