Feature582 (progress) start on writing out vrml files for each vtk model
This commit is a start on writing vrml files for blender.  There is a
method added to the generated python script to load vrml.  I refactored
the code to get the surface and atoms from a model into ModelUtilities
and finally gave ModelUtilities its own file instead of sharing with
SketchModel.  There is a method in ProjectToBlenderAnimation to write out
the file and return its name to be included in the Blender animation.