Feature582 (progress) blender export bugfix and cleanup
This commit fixes a bug I found in the Blender export where the user
cancels the select avi file window and an error message was produced.
I also cleaned up some of the python code generation (it now uses more
functions for repeated code) which reduces the size of the generated
python file by a factor of 3. Still working on choosing how an object
is colored.  Right now it defaults to red.