Feature582 (progress) objects have color mode and array to color by
Added color mode and array to color by attributes to objects.  This
commit makes it possible to color by an arbitrary array with one of
the pre-programmed color maps (currently one color or blue to red).
As with most things, there is no way to get to this from the GUI yet