Feature582 (progress) now exports object colors to blender
The object colors that are selected for the object based on its color
map and array to color by are exported to the vrml files and put into
Blender correctly.  Now I need to add a way to select the colors from the GUI
and to add the new data that is part of an object to the object's saved
data in the project file writer.