Feature722 (progress) added new mode for editing groups
This commit adds a new mode for editing groups.  The only button that
does anything (yet) in this mode is the 'add to group' button, and it
is buggy.  Objects added to a group disappear because the groups' actors
are not updated correctly when objects are added/removed from the group.