Feature722 (progress) Adding objects to group works
This commit also adds the ability to remove objects from groups.  However,
this functionality still has the following issues:  removing an object from
a group makes the group no longer selectable (may only be an issue for groups
of size 2), and there is no indication (other than shadow) that the object
that you are trying to remove from the group is actually selected by the tracker.
So I am not calling remove-from-group done yet.  (They are the same button, it
simply depends on whether or not the trackers are selecting things in the same
group or not what happens).