Feature722 (progress) HUD message updates as objects selected
The message in the HUD for add/remove from group updates to show the
operation that is valid with the current selection.  This also fixes
a bug in ObjectGroup that occurs with only one object in the group.
The vtkTransform segfault happens if you add/remove from the same group
too much though... I think I'll have to track it down and fix it, something
I've been avoiding for a while now.