Feature670/733 (progress) created a vtk source for the keyframe indicator lines
This source the vtkGiftRibbonSource takes the bounds of a box and creates lines that
are similar to the ribbons on a traditionally wrapped gift (thus the name).  Each face
of the box is bisected by two perpendicular lines through the midpoints of opposite sides.
This indicator will be noticable even when the keyframed object is selected since the 'icon'
is different from the hand-selected object 'icon'.