Feature670/733 there is an indicator of keyframes and keyframes can be removed
The 'gift ribbons' from the previous commit are now placed on objects that are
keyframed at the current time to indicate that they are keyframed.  The add keyframe
button now removes the keyframe from an object if the object was keyframed at the current
time and has not moved.  If it has moved, it updates the position of the keyframe instead.