Feature730 (progress) replicas are now in a group
This was started on the way to the SUI conference, but due to failing tests, not
committed.  Replicas and their original objects are now stored in a group. Groups
have been changed to allow objects within them to move (although there is now way
to select the objects to move...).  All tests have been updated to match these new
assumptions.  I also found several bugs in the old code that were revealed when
what was being tested changed.