Feature730 (progress) added ability to select within groups
This includes a refactor of how outlines work that removes a lot of duplicated
code.  Moving the right analog up/down now changes the 'level' within a group
hierarchy that is being examined (in some modes... still working on that).
Grabbing an object within a group still moves the whole group though.