Feature730 (progress) objects within groups can be grabbed individually
There are still some issues with this, but I'm working on them.  This is
a progress commit.  Objects can either get forces themselves or pass the
force to their parent.  Changed the physics to check for forces on objects
within groups and apply them correctly.  Also fixed a bug in structure
replication that I found.

Problems:
Binary collision search mode has some bugs when an object within a group
is grabbed.
Collisions between objects in the same group are not checked yet.