Feature730 (progress) the old-style brute force collision detection sort of works again
That mode doesn't handle collisions between an object within a group that
is being moved and an object that is not in the group well.  But that mode
doesn't have info about which objects were moved in the collision detection
routine so there isn't anything I can do about that.  It works for most cases.
This means that all the collision detection routines work again with the new
assumption that objects within groups are not rigid relative to each other.