Bug810 - fixed a bug in copy where models got confused
It was possible for models and objects to get confused in the old copy-paste
code and even for an object's model to fail to be saved at all based on its
position in the group.  This only affected multi-level groups, which I had not
tested the copy-paste code on before.