Bug812 and Bug 792 - removed corner cases from ObjectGroup
ObjectGroup used to be treated like a single object when there was
only one object inside it.... this led to confusing corner cases that
needed to work different ways in different situations, so I removed this
behavior and everything worked better.  The only thing that broke was the
save code, which is fixed now.  The save tests got a full lookover and
cleanup with this commit too.