TESTING - redid tests of save and load code to make it easier to add new ones
In the process of updating these, I found several bugs, some I committed earlier
and some are in this commit.  There were two potential segfaults and some things
that just didn't behave right.  This also has the valuable change that it is now
trivial to add new save/load tests, and although no new ones are added yet the
current tests (which test the same things as before) are several hundred lines
shorter than they used to be.