BUGFIX - fixed the process for generating simplified models
The simplified model surface was getting appended with the original
surface resulting in something even more complex to render.  Rendering
was already a problem, so this was not helping.  Now the original geometry
is correctly filtered out and the simplified model is the simplified geometry
and the atom data from Chimera.