Feature728 - added a button to set the relative transform between two objects
Currently this pops up a dialog to ask for the transform (in Paraview
translation & euler format).  This doesn't do anything fancy like live
updates yet though.  The dialog is pre-loaded with the current
transformation between the objects.

This commit also rearranges other controls, moving create camera to
the animations mode and visibility settings to the editing color maps
mode.