Feature858/594 - load a model from a locally created pdb file
This commit adds an option under File->Open to open a local PDB file.
There is also a test of this added as well as some modifications to
the Subprocess runner framework.  This also changes the model tested
by several subprocess tests to reduce test time by a factor of 6.