FEATURES - added 'keyframe all' button, and set camera to viewpoint controls
This commit is the features needed to make the video for the reviewer board
on 2013-10-10.  A button to keyframe all objects is added, and menu items
both to create a camera for the current viewpoint, and setting the camera to
the current viewpoint.  There is also a button to delete a spring.  I also
fixed a bug in exporting a keyframed camera to Blender.  The Blender code
had the per-camera lighting parameters changed and now saves a Blender project
file in the project directory.