BUGFIX-fixes edit transform between objects button
This button was using the wrong preconcat/postconcat rule in the matrix
used to calculate the new position and also was ignoring the fact that
the objects could have parents.  These issues are both fixed and the transformations
work correctly even when the objects (and/or their parents) are not at the origin.