Added Hand objects within the project
These keep track of the nearest object and connector and implement
the various grab functions.  This commit also modifies SOME of the
input module to check whether these work.  However the whole of the
input module does not use these since most of it is getting scrapped
anyway.

Due to this, this commit should not be considered a working
version of SketchBio, but a stepping stone for the others in the 523
group to work from.